@@ -6,13 +6,12 @@
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"test": "npm run test:integration && npm run test:unit",
+		"test": "npm run test:integration",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
 		"test:integration": "playwright test",
-		"test:unit": "vitest",
 		"generate": "drizzle-kit generate",
 		"migrate": "drizzle-kit migrate"
 	},
@@ -22,7 +21,7 @@
 		"@skeletonlabs/tw-plugin": "^0.4.0",
 		"@sveltejs/adapter-auto": "^3.2.0",
 		"@sveltejs/kit": "^2.5.8",
-		"@sveltejs/vite-plugin-svelte": "^3.1.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"@types/eslint": "^8.56.10",
 		"@types/node": "^20.12.11",
 		"@typescript-eslint/eslint-plugin": "^6.0.0",
@@ -36,8 +35,8 @@
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.3",
 		"prettier-plugin-tailwindcss": "^0.5.14",
-		"svelte": "^4.2.17",
-		"svelte-check": "^3.7.1",
+		"svelte": "^5.38.10",
+		"svelte-check": "^4.3.1",
 		"tailwindcss": "^3.4.3",
 		"tslib": "^2.6.2",
 		"typescript": "^5.4.5",
